fix(scripts): stop pinning a self-check to an extension version
Bumping a bundled extension's version is mandatory whenever its contents
change, because a device only re-extracts a directory whose name it does
not already have. So a self-check that resolves the extension by a pinned
path breaks on the next correct change, and the version written into it
is stale the moment anyone edits the extension.

test-process-monitor-extension.js was pinned to 1.0.0 and failed with
MODULE_NOT_FOUND on a branch whose only fault was bumping the extension
it tests. test-bridge-relay.js already resolved by prefix and explained
why; that resolver now lives in scripts/lib/bundled-extension.js and both
read it, so the reasoning has one home rather than one copy that rots and
one that does not.

It throws rather than returning nothing when no directory matches. Every
caller is a self-check, and a check that quietly examines nothing reads
exactly like a check that passed.
